+++ src/Transformers/VarianceThresholdFilter.php
@@ -10,12 +10,15 @@ use Rubix\ML\Specifications\SamplesAreCompatibleWithTransformer;
 use Rubix\ML\Exceptions\InvalidArgumentException;
 use Rubix\ML\Exceptions\RuntimeException;
 
+use function Rubix\ML\warn_deprecated;
 use function is_null;
 
 /**
  * Variance Threshold Filter
  *
  * A type of feature selector that selects features with the greatest variance.
+ * 
+ * @deprecated
  *
  * @category    Machine Learning
  * @package     Rubix/ML
@@ -43,6 +46,8 @@ class VarianceThresholdFilter implements Transformer, Stateful, Persistable
      */
     public function __construct(int $minFeatures)
     {
+        warn_deprecated('Variance Threshold Filter is deprecated, use K Best Feature Selector instead.');
+
         if ($minFeatures < 1) {
             throw new InvalidArgumentException('Min features must be'
                 . " greater than 0, $minFeatures given.");
@@ -131,6 +136,6 @@ class VarianceThresholdFilter implements Transformer, Stateful, Persistable
      */
     public function __toString() : string
     {
-        return "Variance Threshold Filter (max_features: {$this->minFeatures})";
+        return "Variance Threshold Filter (min_features: {$this->minFeatures})";
     }
 }